Simplify and speed up the docker build
Use the the 0.12-onbuild image instead of the 0.12 node image to speed up and
simplify the ember-lightning image build. The -onbuild images integrate the
standard node build steps of child images into the parent image in a way that
minimizes unnecessary `npm install` execution and also eliminates the steps
from the derived build.